refactor(cli): address review feedback on outdated-dependency handling
- Centralize outdated-dependency reporting in GenerateService so the `.warn`
  case emits the alert from the same place as `.install` and `.fail`. The
  duplicate `LintingIssue` warning from SwiftPackageManagerGraphLoader is
  removed (use `onOutdatedDependencies: .fail` to fail generation instead
  of `warningsAsErrors: .only([.outdatedDependencies])`).
- Move OutdatedDependenciesChecking/Checker into its own file.
- Leave a comment explaining why the outdated-dependency tests live in their
  own struct (Mockable + MockOpening interaction triggers a Swift Testing
  struct-destroy crash on parallel runs).
