+++ cli/Sources/TuistKit/Services/GenerateService.swift
@@ -1,12 +1,9 @@
-import FileSystem
 import Foundation
-import Mockable
 import Path
 import TuistAlert
 import TuistCache
 import TuistConfig
 import TuistConfigLoader
-import TuistConstants
 import TuistCore
 import TuistEnvironment
 import TuistGenerator
@@ -101,12 +98,14 @@ public struct GenerateService {
             )
 
         if let generatedProject = config.project.generatedProject,
-           generatedProject.generationOptions.onOutdatedDependencies != .warn,
            try await outdatedDependenciesChecker.packageDependenciesAreOutdated(at: path)
         {
             switch generatedProject.generationOptions.onOutdatedDependencies {
             case .warn:
-                break
+                AlertController.current.warning(.alert(
+                    "We detected outdated dependencies.",
+                    takeaway: "Run \(.command("tuist install")) to update them."
+                ))
             case .install:
                 Logger.current.notice("Outdated dependencies detected. Running `tuist install`.", metadata: .section)
                 try await installService.run(
@@ -162,50 +161,5 @@ enum GenerateServiceError: FatalError, Equatable {
         case .outdatedDependencies:
             return "Outdated dependencies detected. Run `tuist install` to update them."
         }
-    }
-}
-
-@Mockable
-protocol OutdatedDependenciesChecking {
-    func packageDependenciesAreOutdated(at path: AbsolutePath) async throws -> Bool
-}
-
-struct OutdatedDependenciesChecker: OutdatedDependenciesChecking {
-    private let manifestFilesLocator: ManifestFilesLocating
-    private let fileSystem: FileSysteming
-
-    init(
-        manifestFilesLocator: ManifestFilesLocating = ManifestFilesLocator(),
-        fileSystem: FileSysteming = FileSystem()
-    ) {
-        self.manifestFilesLocator = manifestFilesLocator
-        self.fileSystem = fileSystem
-    }
-
-    func packageDependenciesAreOutdated(at path: AbsolutePath) async throws -> Bool {
-        guard let packageManifestPath = try await manifestFilesLocator.locatePackageManifest(at: path) else {
-            return false
-        }
-
-        let packageDirectory = packageManifestPath.parentDirectory
-        let currentPackageResolvedPath = packageDirectory
-            .appending(component: Constants.SwiftPackageManager.packageResolvedName)
-        let savedPackageResolvedPath = packageDirectory.appending(components: [
-            Constants.SwiftPackageManager.packageBuildDirectoryName,
-            Constants.DerivedDirectory.name,
-            Constants.SwiftPackageManager.packageResolvedName,
-        ])
-
-        var currentData: Data?
-        if try await fileSystem.exists(currentPackageResolvedPath) {
-            currentData = try await fileSystem.readFile(at: currentPackageResolvedPath)
-        }
-
-        var savedData: Data?
-        if try await fileSystem.exists(savedPackageResolvedPath) {
-            savedData = try await fileSystem.readFile(at: savedPackageResolvedPath)
-        }
-
-        return currentData != savedData
     }
 }
