+++ cli/Tests/TuistLoaderTests/Loaders/SwiftPackageManagerGraphLoaderTests.swift
@@ -454,65 +454,4 @@ struct SwiftPackageManagerGraphLoaderTests {
         )
     }
 
-    @Test
-    func load_warnOutdatedDependencies() async throws {
-        try await withMockedDependencies {
-            try await fileSystem.runInTemporaryDirectory(prefix: UUID().uuidString) {
-                temporaryDirectory in
-                // Given
-                let packageSettings = PackageSettings.test()
-
-                let workspacePath = temporaryDirectory.appending(components: [
-                    ".build", "workspace-state.json",
-                ])
-                try await fileSystem.makeDirectory(at: workspacePath.parentDirectory)
-                try await fileSystem.writeText(
-                    """
-                    {
-                      "object" : {
-                        "artifacts" : [],
-                        "dependencies" : []
-                      }
-                    }
-                    """,
-                    at: workspacePath
-                )
-
-                try await fileSystem.makeDirectory(
-                    at: temporaryDirectory.appending(components: [".build", "Derived"])
-                )
-                let savedPackageResolvedPath = temporaryDirectory.appending(components: [
-                    ".build", "Derived", "Package.resolved",
-                ])
-                let currentPackageResolvedPath = temporaryDirectory.appending(
-                    component: "Package.resolved"
-                )
-                try await fileSystem.writeText("outdated", at: savedPackageResolvedPath)
-                try await fileSystem.touch(currentPackageResolvedPath)
-
-                given(packageInfoMapper)
-                    .resolveExternalDependencies(
-                        path: .any,
-                        packageInfos: .any,
-                        packageToFolder: .any,
-                        packageToTargetsToArtifactPaths: .any,
-                        packageModuleAliases: .any,
-                        packageSettings: .any
-                    )
-                    .willReturn([:])
-
-                // When
-                let (_, lintingIssues) = try await subject.load(
-                    packagePath: temporaryDirectory.appending(component: "Package.swift"),
-                    packageSettings: packageSettings,
-                    disableSandbox: true
-                )
-
-                // Then
-                #expect(
-                    lintingIssues.contains { $0.reason.contains("We detected outdated dependencies") }
-                )
-            }
-        }
-    }
 }
